+++ lilio/traintest.py
@@ -5,6 +5,7 @@ Wrapper around sklearn splitters for working with (multiple) xarray dataarrays.
 from collections.abc import Iterable
 from typing import Optional
 from typing import Union
+from typing import overload
 import numpy as np
 import xarray as xr
 from sklearn.model_selection._split import BaseCrossValidator
@@ -12,13 +13,7 @@ from sklearn.model_selection._split import BaseShuffleSplit
 
 
 # Mypy type aliases
-XType = Union[xr.DataArray, list[xr.DataArray]]
 CVtype = Union[BaseCrossValidator, BaseShuffleSplit]
-
-# For output types, variables are split in 2
-XOnly = tuple[XType, XType]
-XAndY = tuple[XType, XType, xr.DataArray, xr.DataArray]
-XMaybeY = Iterable[Union[XOnly, XAndY]]
 
 
 class CoordinateMismatchError(Exception):
@@ -55,28 +50,106 @@ class TrainTestSplit:
         """
         self.splitter = splitter
 
+    @overload
     def split(
         self,
-        *x_args: xr.DataArray,
+        x_args: xr.DataArray,
         y: Optional[xr.DataArray] = None,
         dim: str = "anchor_year",
-    ) -> XMaybeY:
+    ) -> Iterable[tuple[xr.DataArray, xr.DataArray, xr.DataArray, xr.DataArray]]:
+        ...
+
+    @overload
+    def split(
+        self,
+        x_args: Iterable[xr.DataArray],
+        y: Optional[xr.DataArray] = None,
+        dim: str = "anchor_year",
+    ) -> Iterable[
+        tuple[
+            Iterable[xr.DataArray], Iterable[xr.DataArray], xr.DataArray, xr.DataArray
+        ]
+    ]:
+        ...
+
+    def split(
+        self,
+        x_args: Union[xr.DataArray, Iterable[xr.DataArray]],
+        y: Optional[xr.DataArray] = None,
+        dim: str = "anchor_year",
+    ) -> Iterable[
+        Union[
+            tuple[xr.DataArray, xr.DataArray],
+            tuple[xr.DataArray, xr.DataArray, xr.DataArray, xr.DataArray],
+            tuple[Iterable[xr.DataArray], Iterable[xr.DataArray]],
+            tuple[
+                Iterable[xr.DataArray],
+                Iterable[xr.DataArray],
+                xr.DataArray,
+                xr.DataArray,
+            ],
+        ]
+    ]:
         """Iterate over splits.
 
         Args:
             x_args: one or multiple xr.DataArray's that share the same
-                coordinate along the given dimension
+                coordinate along the given dimension.
             y: (optional) xr.DataArray that shares the same coordinate along the
-                given dimension
+                given dimension.
             dim: name of the dimension along which to split the data.
 
         Returns:
             Iterator over the splits
         """
+        x_args_list, x = self._check_dimension_and_type(x_args, y, dim)
+
+        # Now we know that all inputs are equal.
+        for train_indices, test_indices in self.splitter.split(x[dim]):
+            x_train = [da.isel({dim: train_indices}) for da in x_args_list]
+            x_test = [da.isel({dim: test_indices}) for da in x_args_list]
+
+            if y is None:
+                if isinstance(x_args, xr.DataArray):
+                    yield x_train.pop(), x_test.pop()
+                else:
+                    yield x_train, x_test
+            else:
+                y_train = y.isel({dim: train_indices})
+                y_test = y.isel({dim: test_indices})
+                if isinstance(x_args, xr.DataArray):
+                    yield x_train.pop(), x_test.pop(), y_train, y_test
+                else:
+                    yield x_train, x_test, y_train, y_test
+
+    def _check_dimension_and_type(
+        self,
+        x_args: Union[xr.DataArray, Iterable[xr.DataArray]],
+        y: Optional[xr.DataArray] = None,
+        dim: str = "anchor_year",
+    ) -> tuple[list[xr.DataArray], xr.DataArray]:
+        """Check input dimensions and type and return input as list.
+
+        Args:
+            x_args: one or multiple xr.DataArray's that share the same
+                coordinate along the given dimension.
+            y: (optional) xr.DataArray that shares the same coordinate along the
+                given dimension.
+            dim: name of the dimension along which to split the data.
+
+        Returns:
+            List of input x and dataarray containing coordinate info
+        """
         # Check that all inputs share the same dim coordinate
         coords = []
         x: xr.DataArray  # Initialize x to set scope outside loop
-        for x in x_args:
+
+        if isinstance(x_args, xr.DataArray):
+            x_args_list = [x_args]
+        else:
+            x_args_list = list(x_args)
+
+        for x in x_args_list:
             try:
                 coords.append(x[dim])
             except KeyError as err:
@@ -96,21 +169,7 @@ class TrainTestSplit:
 
         if x[dim].size <= 1:
             raise ValueError(
-                f"Invalid input: need at least 2 values along dimension {dim}"
+                f"Invalid input: need at least 2 values along dimension {dim}."
             )
 
-        # Now we know that all inputs are equal.
-        for train_indices, test_indices in self.splitter.split(x[dim]):
-            if len(x_args) == 1:
-                x_train: XType = x.isel({dim: train_indices})
-                x_test: XType = x.isel({dim: test_indices})
-            else:
-                x_train = [da.isel({dim: train_indices}) for da in x_args]
-                x_test = [da.isel({dim: test_indices}) for da in x_args]
-
-            if y is None:
-                yield x_train, x_test
-            else:
-                y_train = y.isel({dim: train_indices})
-                y_test = y.isel({dim: test_indices})
-                yield x_train, x_test, y_train, y_test
+        return x_args_list, x