+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+- Consistent output type of train-test split as input ([#62](https://github.com/AI4S2S/lilio/pull/62)).
+
 ## 0.4.1 (2023-09-11)
 ### Added
 - Python 3.11 support ([#60](https://github.com/AI4S2S/lilio/pull/60)).